# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Expo AHRO Telecoms",
        "description": "Documentación de la API para Expo AHRO",
        "contact": {
            "email": "tu-email@dominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "operationId": "70a301af8ea04ce2f2596a58758632cd",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Agregar un nuevo producto",
                "operationId": "4ff8851cb6bfe92aa226cdc804917ff0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del producto",
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "description": "Descripción del producto",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "operationId": "20275afdbab1759ad2d46f3dcfbeaeeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "ea733033fbff1681e04f8c4d0e11e0bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
        "/api/seguimientos": {
            "post": {
                "tags": [
                    "Seguimientos"
                ],
                "summary": "Registrar un nuevo seguimiento",
                "operationId": "8a5b548eef49390190e47bbf949ceaf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitante_id",
                                    "tipo",
                                    "comentarios"
                                ],
                                "properties": {
                                    "visitante_id": {
                                        "type": "integer"
                                    },
                                    "tipo": {
                                        "description": "Tipo de seguimiento",
                                        "type": "string"
                                    },
                                    "comentarios": {
                                        "description": "Comentarios",
                                        "type": "string"
                                    },
                                    "fecha": {
                                        "description": "Fecha del seguimiento",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seguimiento creado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/seguimientos/{id}": {
            "put": {
                "tags": [
                    "Seguimientos"
                ],
                "summary": "Actualizar un seguimiento",
                "operationId": "8e2fc546a437fda83d24e1c25f6b07c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del seguimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "comentarios": {
                                        "type": "string"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seguimiento actualizado"
                    },
                    "404": {
                        "description": "Seguimiento no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Seguimientos"
                ],
                "summary": "Eliminar un seguimiento",
                "operationId": "ca98e86fa50b114bac0794354c1e0c22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del seguimiento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seguimiento eliminado"
                    },
                    "404": {
                        "description": "Seguimiento no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Producto": {
                "title": "Producto",
                "required": [
                    "id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Producto A"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción del producto A"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}